+++ src/app/search/search-filters/loadouts.tsx
@@ -1,4 +1,5 @@
 import { tl } from 'app/i18next-t';
+import { getHashtagsFromNote } from 'app/inventory/note-hashtags';
 import { Loadout } from 'app/loadout-drawer/loadout-types';
 import { FilterDefinition } from '../filter-types';
 import { quoteFilterString } from '../query-parser';
@@ -25,8 +26,14 @@ const loadoutFilters: FilterDefinition[] = [
       // a search like
       // inloadout:"loadout name here"
       // inloadout:"pvp" (for all loadouts with pvp in their name)
+      // inloadout:"#pve" (for all loadouts with the #pve hashtag in name or notes)
       return (item) =>
-        loadoutsByItem[item.id]?.some((l) => l.loadout.name.toLowerCase().includes(filterValue));
+        loadoutsByItem[item.id]?.some(
+          (l) =>
+            l.loadout.name.toLowerCase().includes(filterValue) ||
+            (filterValue.startsWith('#') && // short circuit for less load
+              getHashtagsFromNote(l.loadout.notes).includes(filterValue))
+        );
     },
   },
 ];